+++ packages/dnd_kit_flutter/lib/dnd_kit_flutter.dart
@@ -25,7 +25,6 @@ library;
 
 export 'package:dnd_kit/dnd_kit.dart';
 
-export 'src/a11y/announcements.dart';
 export 'src/measuring/measuring.dart' hide DndMeasuredBox;
 export 'src/scope/controller.dart';
 export 'src/scope/scope.dart';